+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import logo from './logo.svg';
 import './App.css';
-import Dashboard from '../src/components/Dashboard';
-import L1eval from '../src/components/L1eval';
+import Dashboard from '../src/components/Dashboard/Dashboard'
+import L1eval from '../src/components/L1eval/L1eval';
 import Gkeval from '../src/components/Gkeval';
 
 class App extends Component {
