+++ src/components/Dashboard/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Component } from 'react';
 import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
-import '../../node_modules/react-bootstrap-table/dist/react-bootstrap-table-all.min.css';
+import '../../../node_modules/react-bootstrap-table/dist/react-bootstrap-table-all.min.css';
 
 class Dashboard extends Component {
   constructor(props) {
