--- src/components/InputText.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Component } from 'react';
-
-
-const InputText = (props) => (  
-  <div className="form-group">
-    {/* <input
-      className={"col-"+props.className}
-      type={'text'}
-      placeholder={props.placeholder} /> */}
-  </div>
-);
-
-export default InputText;
