+++ src/components/L1eval/L1eval.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import { Component } from 'react';
-import InputText from './resuableComponent/InputText';
-import InputSelect from './resuableComponent/InputSelect';
-import InputTextArea from './resuableComponent/InputTextArea';
-import Label from './resuableComponent/Label';
-// import DropDown from './resuableComponent/DropDown';
-import InputNumber from './resuableComponent/InputNumber';
-import Button from './resuableComponent/Button';
-import InputFile from './resuableComponent/InputFile';
+import InputText from '../resuableComponent/InputText';
+import InputSelect from '../resuableComponent/InputSelect';
+import InputTextArea from '../resuableComponent/InputTextArea';
+import Label from '../resuableComponent/Label';
+import InputNumber from '../resuableComponent/InputNumber';
+import Button from '../resuableComponent/Button';
+import InputFile from '../resuableComponent/InputFile';
 
 class L1eval extends Component {
   constructor(props) {
