--- src/components/TextBox.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Component } from 'react';
-
-const TextBox = (props) => (
-  <div>
-    <input  type="text" className="form-control"/>
-  </div>
-);
-
-
-export default TextBox;
